% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DriverRWH_score.R
\name{get_DriverRWH_score}
\alias{get_DriverRWH_score}
\title{get_DriverRWH_score}
\usage{
get_DriverRWH_score(Mutation, string, theta = 0.2, out_file_dir)
}
\arguments{
\item{Mutation}{the mutation data, rows represent genes and columns represent samples}

\item{string}{network data, it should have at least two columns and the first two elements of each row are considered two vertices of an unweighted edge. Other elements in the row will be omitted}

\item{theta}{the restart probability at every step of the random walk}

\item{out_file_dir}{the directory where you wish to put result of the DriverRWH}
}
\description{
get_DriverRWH_score uses the mutation data, network data to perform DriverRWH.
theta is the restart probability at every step of the random walk.
result of the DriverRWH will be in out_file_dir
}
\examples{

###the example mutation data is a Lung Squamous Cell Carcinoma Dataset from TCGA
###the example network data is HumanNet
###We will show how to get DriverRWH Results

library(DriverRWH)

########## load the example mutation data ##########
#you can also use your own mutation data
#rows represent genes and columns represent samples
#rownames and colnames of mutation data cannot be omitted
data(luscExampleMutation)

##########     load the network data      ##########
#you can use HumanNet or STRINGv10 by loading HumanNet.rda or STRINGv10.rda
#you can also use your own network data
#the network data should have at least two columns
#the first two elements of each row are considered two vertices of an unweighted edge.
#Other elements in the row will be omitted
data(HumanNet)

#### set the current working directory as the output file directory ####
my_file_dir=getwd()

##########   get the result of DriverRWH   ##########
#the output file would be in the out_data folder in the output file directory
#the default value of theta is 0.2
get_DriverRWH_score(Mutation=luscExampleMutation,string=HumanNet,out_file_dir=my_file_dir)

}
