% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P.R
\name{get_P}
\alias{get_P}
\title{get_P}
\usage{
get_P(Mutation, vertex_W)
}
\arguments{
\item{Mutation}{the mutation data, rows represent genes and columns represent samples}

\item{vertex_W}{the vertex weight of the hypergraph}
}
\description{
get_P uses the mutation data and the vertex weight of the hypergraph to calculate the transition possibility matrix
output of get_P is the transition possibility matrix of the hypergraph
}
\examples{

###the vertex weight of the hypergraph is calculated by get_vertex_W(Mutation,graph)
###and some relevant data should be calculated

###the example mutation data is a Lung Squamous Cell Carcinoma Dataset from TCGA
###We will show how to get the transition possibility matrix of the hypergraph

library(DriverRWH)

#load the mutation data
data(luscExampleMutation)

#load the network data
data(HumanNet)

#get the graph from HumanNet
graph=graph.data.frame(HumanNet)

#get the vertex weight of the hypergraph
W=get_vertex_W(Mutation=luscExampleMutation,graph=graph)

#get the transition possibility matrix P
P=get_P(Mutation=luscExampleMutation,vertex_W=W)

}
