% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_W.R
\name{get_vertex_W}
\alias{get_vertex_W}
\title{get_vertex_W}
\usage{
get_vertex_W(Mutation, graph)
}
\arguments{
\item{Mutation}{the mutation data, rows represent genes and columns represent samples}

\item{graph}{the graph constructed from network data}
}
\description{
get_vertex_W uses the mutation data and the graph constructed from network data to calculate the weight of the hypergraph
output of get_vertex_W is the vertex weight of the hypergraph
}
\examples{

###the example mutation data is a Lung Squamous Cell Carcinoma Dataset from TCGA
###the example graph is constructed from the HumanNet
###We will show how to get the vertex weight of the hypergraph

library(DriverRWH)

#load the mutation data
data(luscExampleMutation)

#load the network data
data(HumanNet)

#get the graph from HumanNet
graph=graph.data.frame(HumanNet)

#or from your own network
#network_file=""
#string=read.table(file=network_file,header=TRUE)
#graph=graph.data.frame(string)

#get the vertex weight of the hypergraph
W=get_vertex_W(Mutation=luscExampleMutation,graph=graph)

}
