% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(mutation_file)
}
\arguments{
\item{mutation_file}{the text that contains the mutation data, a gene*sample matrix with rownames being gene names and column names being sample IDs}
}
\value{
Mutation, the mutation data, rows represent genes and columns represent samples
}
\description{
read_data uses reads the mutation file
it deletes some row(s) of the mutation data whose rowsum is zero, rownames of the remaining rows are considered possibale genes
}
\examples{

###read_data() reads and processes the mutation data

###it deletes some row(s) of the mutation data whose rowsum is zero,
###rownames of the remaining rows are considered possibale genes

library(DriverRWH)

###place the mutation file in the current working directory and name it "mutation_file"
###to run the example code, please uncomment them
#my_file_dir=getwd()
#mutation_file=paste0(my_file_dir,"/mutation_file.txt")

#Mutation=read_data(mutation_file=mutation_file)

###the mutation data
#print(Mutation)


}
